requirements.txt
flask
python-dotenv
google-genai
PyPDF2
python-docx
reportlab
flask-limiter